refactor(oidc): address PR review — merge services, simplify cache, rename files
- Merge smtp-config + smtp-admin → smtp.ts (resolver + CRUD in one file)
- Merge social-config + social-admin → social.ts (same)
- Inline ENCRYPTION_KEY_VERSION per file, drop encryption-key-version.ts
- Drop in-flight promise dedup from ttl-cache (YAGNI for admin-reconfigured data)
- Rename realm-resolver.ts → oidc-realm-resolver.ts (naming consistency)
- Add client-side Zod validation in use-app-auth-config.ts (port range, email)
- Document AsyncLocalStorage propagation invariant in social override plugin
- Net: -77 lines, same behavior, all tests green
