+++ apps/api/src/modules/oidc/test/integration/routes/oauth-email-verification.test.ts
@@ -45,8 +45,7 @@ import { truncateAll } from "../../../../../../test/helpers/db.ts";
 import { enableSmtpForSuite } from "../../../../../../test/helpers/smtp.ts";
 import { createClient } from "../../../services/oauth-admin.ts";
 import { _resetClientCache } from "../../../services/oauth-admin.ts";
-import { upsertSmtpConfig } from "../../../services/smtp-admin.ts";
-import { _clearSmtpCacheForTesting } from "../../../services/smtp-config.ts";
+import { upsertSmtpConfig, _clearSmtpCacheForTesting } from "../../../services/smtp.ts";
 import oidcModule from "../../../index.ts";
 
 const app = getTestApp({ modules: [oidcModule] });
