+++ apps/web/src/modules/oidc/hooks/use-app-auth-config.ts
@@ -4,9 +4,14 @@
  * React Query hooks for per-application SMTP + social auth configuration.
  * Mirrors `/api/applications/:id/smtp-config` and
  * `/api/applications/:id/social-providers/:provider` routes.
+ *
+ * Inputs are Zod-validated client-side before hitting the API so the form
+ * gets immediate feedback on obvious errors (bad port range, malformed
+ * email). The server still validates — this is purely UX.
  */
 
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
+import { z } from "zod";
 import { api, ApiError } from "@/api";
 import { useCurrentOrgId } from "@/hooks/use-org";
 import { useCurrentApplicationId } from "@/hooks/use-current-application";
@@ -34,6 +39,26 @@ export interface SocialProviderView {
   updatedAt: string;
 }
 
+export const upsertSmtpSchema = z.object({
+  host: z.string().min(1, "Host is required"),
+  port: z.number().int().min(1).max(65535),
+  username: z.string().min(1, "Username is required"),
+  pass: z.string().min(1, "Password is required"),
+  fromAddress: z.email("Invalid email address"),
+  fromName: z.string().nullable().optional(),
+  secureMode: z.enum(["auto", "tls", "starttls", "none"]).optional(),
+});
+
+export type UpsertSmtpInput = z.infer<typeof upsertSmtpSchema>;
+
+export const upsertSocialSchema = z.object({
+  clientId: z.string().min(1, "Client ID is required"),
+  clientSecret: z.string().min(1, "Client secret is required"),
+  scopes: z.array(z.string()).nullable().optional(),
+});
+
+export type UpsertSocialInput = z.infer<typeof upsertSocialSchema>;
+
 async function getOrNull<T>(path: string): Promise<T | null> {
   try {
     return await api<T>(path);
@@ -57,19 +82,13 @@ export function useUpsertSmtpConfig() {
   const qc = useQueryClient();
   const appId = useCurrentApplicationId();
   return useMutation({
-    mutationFn: (data: {
-      host: string;
-      port: number;
-      username: string;
-      pass: string;
-      fromAddress: string;
-      fromName?: string | null;
-      secureMode?: "auto" | "tls" | "starttls" | "none";
-    }) =>
-      api<SmtpConfigView>(`/applications/${appId}/smtp-config`, {
+    mutationFn: (data: UpsertSmtpInput) => {
+      const parsed = upsertSmtpSchema.parse(data);
+      return api<SmtpConfigView>(`/applications/${appId}/smtp-config`, {
         method: "PUT",
-        body: JSON.stringify(data),
-      }),
+        body: JSON.stringify(parsed),
+      });
+    },
     onSuccess: () => {
       qc.invalidateQueries({ queryKey: ["smtp-config"] });
     },
@@ -89,12 +108,15 @@ export function useDeleteSmtpConfig() {
 
 export function useTestSmtp() {
   const appId = useCurrentApplicationId();
+  const toSchema = z.email("Invalid email address");
   return useMutation({
-    mutationFn: (to: string) =>
-      api<{ ok: boolean; messageId?: string; error?: string }>(
+    mutationFn: (to: string) => {
+      const parsed = toSchema.parse(to);
+      return api<{ ok: boolean; messageId?: string; error?: string }>(
         `/applications/${appId}/smtp-config/test`,
-        { method: "POST", body: JSON.stringify({ to }) },
-      ),
+        { method: "POST", body: JSON.stringify({ to: parsed }) },
+      );
+    },
   });
 }
 
@@ -113,11 +135,13 @@ export function useUpsertSocialProvider(provider: SocialProviderId) {
   const qc = useQueryClient();
   const appId = useCurrentApplicationId();
   return useMutation({
-    mutationFn: (data: { clientId: string; clientSecret: string; scopes?: string[] | null }) =>
-      api<SocialProviderView>(`/applications/${appId}/social-providers/${provider}`, {
+    mutationFn: (data: UpsertSocialInput) => {
+      const parsed = upsertSocialSchema.parse(data);
+      return api<SocialProviderView>(`/applications/${appId}/social-providers/${provider}`, {
         method: "PUT",
-        body: JSON.stringify(data),
-      }),
+        body: JSON.stringify(parsed),
+      });
+    },
     onSuccess: () => {
       qc.invalidateQueries({ queryKey: ["social-provider"] });
     },
